+++ model_column.go
@@ -24,9 +24,9 @@ package asposecellscloud
 
 type Column struct {
 	Link *Link `json:"link,omitempty" xml:"link"`
-	Index int32 `json:"Index,omitempty" xml:"Index"`
+	Index int64 `json:"Index,omitempty" xml:"Index"`
 	Style *LinkElement `json:"Style,omitempty" xml:"Style"`
-	GroupLevel int32 `json:"GroupLevel,omitempty" xml:"GroupLevel"`
+	GroupLevel int64 `json:"GroupLevel,omitempty" xml:"GroupLevel"`
 	IsHidden bool `json:"IsHidden,omitempty" xml:"IsHidden"`
 	Width float64 `json:"Width,omitempty" xml:"Width"`
 }
