+++ model_font.go
@@ -32,5 +32,5 @@ type Font struct {
 	IsSuperscript bool `json:"IsSuperscript,omitempty" xml:"IsSuperscript"`
 	IsStrikeout bool `json:"IsStrikeout,omitempty" xml:"IsStrikeout"`
 	Underline string `json:"Underline,omitempty" xml:"Underline"`
-	Size int32 `json:"Size,omitempty" xml:"Size"`
+	Size int64 `json:"Size,omitempty" xml:"Size"`
 }
