+++ model_image_save_options.go
@@ -37,14 +37,14 @@ type ImageSaveOptions struct {
 	SortNames bool `json:"SortNames,omitempty" xml:"SortNames"`
 	ChartImageType string `json:"ChartImageType,omitempty" xml:"ChartImageType"`
 	EmbededImageNameInSvg string `json:"EmbededImageNameInSvg,omitempty" xml:"EmbededImageNameInSvg"`
-	HorizontalResolution int32 `json:"HorizontalResolution,omitempty" xml:"HorizontalResolution"`
+	HorizontalResolution int64 `json:"HorizontalResolution,omitempty" xml:"HorizontalResolution"`
 	ImageFormat string `json:"ImageFormat,omitempty" xml:"ImageFormat"`
 	IsCellAutoFit bool `json:"IsCellAutoFit,omitempty" xml:"IsCellAutoFit"`
 	OnePagePerSheet bool `json:"OnePagePerSheet,omitempty" xml:"OnePagePerSheet"`
 	OnlyArea bool `json:"OnlyArea,omitempty" xml:"OnlyArea"`
 	PrintingPage string `json:"PrintingPage,omitempty" xml:"PrintingPage"`
-	PrintWithStatusDialog int32 `json:"PrintWithStatusDialog,omitempty" xml:"PrintWithStatusDialog"`
-	Quality int32 `json:"Quality,omitempty" xml:"Quality"`
+	PrintWithStatusDialog int64 `json:"PrintWithStatusDialog,omitempty" xml:"PrintWithStatusDialog"`
+	Quality int64 `json:"Quality,omitempty" xml:"Quality"`
 	TiffCompression string `json:"TiffCompression,omitempty" xml:"TiffCompression"`
-	VerticalResolution int32 `json:"VerticalResolution,omitempty" xml:"VerticalResolution"`
+	VerticalResolution int64 `json:"VerticalResolution,omitempty" xml:"VerticalResolution"`
 }
