+++ model_import_csv_data_option.go
@@ -28,9 +28,9 @@ type ImportCsvDataOption struct {
 	DestinationWorksheet  string `json:"DestinationWorksheet,omitempty" xml:"DestinationWorksheet"`
 	IsInsert  bool `json:"IsInsert,omitempty" xml:"IsInsert"`
 	ConvertNumericData  bool `json:"ConvertNumericData,omitempty" xml:"ConvertNumericData"`
-	FirstColumn  int32 `json:"FirstColumn,omitempty" xml:"FirstColumn"`
+	FirstColumn  int64 `json:"FirstColumn,omitempty" xml:"FirstColumn"`
 	SourceFile  string `json:"SourceFile,omitempty" xml:"SourceFile"`
-	FirstRow  int32 `json:"FirstRow,omitempty" xml:"FirstRow"`
+	FirstRow  int64 `json:"FirstRow,omitempty" xml:"FirstRow"`
 	SeparatorString  string `json:"SeparatorString,omitempty" xml:"SeparatorString"`
 	CustomParsers  []CustomParserConfig `json:"CustomParsers,omitempty" xml:"CustomParsers"`
 }
