+++ model_page_break_operate_parameter.go
@@ -24,10 +24,10 @@ package asposecellscloud
 
 type PageBreakOperateParameter struct {
 	OperateType  string `json:"OperateType,omitempty" xml:"OperateType"`
-	Index  int32 `json:"Index,omitempty" xml:"Index"`
-	EndIndex  int32 `json:"EndIndex,omitempty" xml:"EndIndex"`
-	Column  int32 `json:"Column,omitempty" xml:"Column"`
-	StartIndex  int32 `json:"StartIndex,omitempty" xml:"StartIndex"`
+	Index  int64 `json:"Index,omitempty" xml:"Index"`
+	EndIndex  int64 `json:"EndIndex,omitempty" xml:"EndIndex"`
+	Column  int64 `json:"Column,omitempty" xml:"Column"`
+	StartIndex  int64 `json:"StartIndex,omitempty" xml:"StartIndex"`
 	PageBreakType  string `json:"PageBreakType,omitempty" xml:"PageBreakType"`
-	Row  int32 `json:"Row,omitempty" xml:"Row"`
+	Row  int64 `json:"Row,omitempty" xml:"Row"`
 }
